+++ G474/G474_05_adc_periodic_irq/BSP/Src/bsp.c
@@ -47,7 +47,7 @@ static void PeriodicProcess(void)
 {
 	uint8_t keyStatus = KEY_Read();
 	
-	static uint16_t timeCnt = 0;
+	static uint32_t timeCnt = 0;
 	if (++timeCnt < 10) return;
 	timeCnt = 0;
 	
@@ -56,7 +56,7 @@ static void PeriodicProcess(void)
 
 static void SampleProcess(void)
 {
-	static uint16_t timeCnt = 0;
+	static uint32_t timeCnt = 0;
 	if (++timeCnt < 2000) return;
 	timeCnt = 0;
